Fix bugs, handle `@use`s
Fix overall dataflow, making it easier to see the page content flow from
one function to the next.

We now integrate the `@use` directive, which allows to use another page
as a template (instead of merely including its content). This allows a
template to "wrap around" content.

The `@use` and `@include` directives are now being resolved recursively.