Implement page caching
In order to avoid deep recursion when using the same templates,
`mrbones` now caches the result of page generation by default, allowing
subsequent template accesses to be fast.

If caching is not wanted for some reason, it can be disabled through
`--no-cache`.

No-cache baselines are added to the testsuite, to compare `mrbones`'
behavior with and without caching.

This addresses [#3].